refactor: Moved MessageConfig to config package
